refactor: change namespace of IValueCheckers
