+++ src/CuiLib/Options/SingleValueOption.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using CuiLib.Checkers;
 
 namespace CuiLib.Options
 {
