+++ src/CuiLib/_Common/ThrowHelpers.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
 using CuiLib.Options;
+using CuiLib.Checkers;
 
 namespace CuiLib
 {
